Add separate source files for anomaly detection charts.
